news_orchestrator/requirements.txt
streamlit
requests
python-dateutil
dateparser
spacy
sentence-transformers
openai
python-dotenv